Introduce mainnet_test.py
